requirements.txt
# Python == 3.11.3
Django == 4.2.3
psycopg2 == 2.9.6
djangorestframework == 3.14.0
markdown == 3.4.3
django-filter == 23.2
djangorestframework_simplejwt == 5.2.2
django-model-utils == 4.3.1
django-cors-headers == 4.2.0
django-filter